+++ src/main/java/com/unstoppabledomains/exceptions/ns/NSExceptionCode.java
@@ -11,5 +11,6 @@ public enum NSExceptionCode {
   IncorrectMethodName,
   UnspecifiedResolver,
   UnsupportedCurrency,
-  NotImplemented;
+  NotImplemented,
+  InvalidDomain;
 }